requirements.txt
aniso8601==3.0.2
Django==2.0.6
django-cors-headers==2.2.0
django-filter==1.1.0
django-guardian==1.4.9
django-rest-framework==0.1.0
djangorestframework==3.8.2
graphene==2.1.2
graphene-django==2.1rc1
graphene-django-extras==0.3.7
graphql-core==2.1rc2
graphql-relay==0.4.5
iso8601==0.1.12
promise==2.1
pytz==2018.4
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
typing==3.6.4
